Pulled changes from master, refactored code
